resort some test file.
