+++ examples/language_model/glm/tests/glm_mp.py
@@ -16,9 +16,6 @@ import paddle
 
 from paddlenlp.transformers import GLMModel
 
-# from modeling import GLMModel
-
-
 tensor_parallel_degree = paddle.distributed.get_world_size()
 tensor_parallel_rank = paddle.distributed.get_rank()
 strategy = paddle.distributed.fleet.DistributedStrategy()
@@ -30,34 +27,23 @@ strategy.hybrid_configs = {
 }
 paddle.distributed.fleet.init(is_collective=True, strategy=strategy)
 
-# model = AutoModel.from_pretrained("THUDM/glm-large-chinese",
-# config=GLMConfig.from_pretrained("glm-large-chinese")
-# config.mp_degree=mp_degree
-# config.tensor_parallel_rank=tensor_parallel_rank
 
-# from modeling import GLMModel;
+# def func(self, *args, **kwargs):
+#     return
 
-
-def distributed_independent_guard():
-    world_size = paddle.distributed.get_world_size()
-    index = paddle.distributed.get_rank()
-    for i in range(world_size):
-        if i == index:
-            pass
-            paddle.barrier()
-        else:
-            paddle.barrier()
-
+# # 屏蔽init_weights
+# GLMModel.init_weights = func
+# paddle.set_default_dtype("float16")
 
 model = GLMModel.from_pretrained(
     "THUDM/glm-large-chinese",
-    # from_hf_hub=True,
     tensor_parallel_degree=tensor_parallel_degree,
     tensor_parallel_rank=tensor_parallel_rank,
+    # dtype="float16",
+    low_cpu_mem_usage=True,
 )
 
 model.eval()
 ret = model(input_ids=paddle.arange(100, 110, dtype="int64").reshape([1, -1]))
 
-# torch 2.1089835166931152
 print("paddle mp", ret.logits.abs().mean().item())
